Rename Balance.SqlQuery2 to Balance.SqlQuery and remove the old one
